spi updates, add mock-idf files, update test files
