+++ mock-idf/components/freertos/include/freertos/FreeRTOS.h
@@ -1,1 +1,4 @@
 #pragma once
+
+/* Basic FreeRTOS definitions. */
+#include "projdefs.h"
