+++ test/i2c_bme280_test/i2c_bme280_test.cpp
@@ -13,22 +13,17 @@ limitations under the License.
 */
 
 /****************************************************************************************
- * Typical results from running program on esp32-devkit v1
+ * Typical results from running program esp32-DevKit-v1
  *
  * W (74246) APP:: ============ APP TICK TICK  =============
  * I (74247) MemStat: Mem type |  8-bit free | Smallest block | Minimum free | 32-bit free | Smallest block | Minimum free
- * I (74253) MemStat: INTERNAL |      215796 |         113804 |       215448 |      273088 |         113804 |       272732
- * I (74265) MemStat:      DMA |      215796 |         113804 |       215448 |      215796 |         113804 |       215448
+ * I (74253) MemStat: INTERNAL |      216840 |         113804 |       216556 |      273556 |         113804 |       273264
+ * I (74265) MemStat:      DMA |      216840 |         113804 |       216556 |      216840 |         113804 |       216556
  * I (74276) MemStat:   SPIRAM |           0 |              0 |            0 |           0 |              0 |            0
  * I (74287) MemStat:
  * I (74290) MemStat:             Name |      Stack |  Min free stack |  Max used stack
- * I (74299) MemStat:         MainTask |      16384 |           12076 |            4308
- * I (74307) MemStat: SocketDispatcher |      20480 |           18380 |            2100
- * I (74315) APP:: ........................................
- * E (74322) APP:: cmd = 0x09  param [0] = 0xa4
- * E (74326) APP:: cmd = 0x09  param [1] = 0x53
- * E (74331) APP:: cmd = 0x09  param [2] = 0x04
- * E (74336) APP:: cmd = 0x09  param [3] = 0x00
+ * I (74299) MemStat:         MainTask |      16384 |           12084 |            4300
+ * I (74307) MemStat: SocketDispatcher |      20480 |           18388 |            2092
  * I (74341) APP:: ........................................
  * E (74348) APP:: BME280 Temperature  (degC)  = 22.4
  * E (74353) APP:: BME280 Humidity     (%RH)   = 37.0
@@ -43,42 +38,29 @@ limitations under the License.
  * E (74406) APP:: Filter configuration        = FC_OFF
  * E (74412) APP:: Spi Interface configuration = SPI_4_WIRE
  * I (74417) APP:: ........................................
- * E (74424) APP:: BME280 ID  = 0x60
- * I (74427) APP:: ........................................
  ****************************************************************************************/
 #include <vector>
-#include "test_multiple_4_line_devices.h"
+#include "i2c_bme280_test.h"
 #include "smooth/core/logging/log.h"
 #include "smooth/core/SystemStatistics.h"
 
+#include <stdint.h>
+
 using namespace smooth::core;
 using namespace std::chrono;
-using namespace smooth::core::io::spi;
-using namespace smooth::application::display;
 using namespace smooth::application::sensor;
 
-namespace test_multiple_4_line_devices
+namespace i2c_bme280_test
 {
     static const char* TAG = "APP";
-    static const uint8_t READ_ID_INFO = 0x04;
-    static const uint8_t READ_DISPLAY_STATUS = 0x09;
-    static const uint8_t READ_DISPLAY_PWR_MODE = 0x0A;
-    static const uint8_t READ_DISPLAY_MADCTL = 0x0B;
-    static const uint8_t ONE_PARAM = 0x01;
-    static const uint8_t TWO_PARAMS = 0x02;
-    static const uint8_t THREE_PARAMS = 0x03;
-    static const uint8_t FOUR_PARAMS = 0x04;
 
     App::App() : Application(APPLICATION_BASE_PRIO, seconds(3)),
-            spi_host(VSPI_HOST),            // Use VSPI as host
-
-            spi_master(
-                spi_host,                   // host VSPI
-                DMA_1,                      // use DMA
-                GPIO_NUM_23,                // mosi gpio pin
-                GPIO_NUM_19,                // miso gpio pin  (full duplex)
-                GPIO_NUM_18,                // clock gpio pin
-                0)                          // max transfer size default of 4096
+            i2c_master(I2C_NUM_0,                       // I2C Port 0
+                       GPIO_NUM_22,                     // SCL pin
+                       false,                           // SCL internal pullup NOT enabled
+                       GPIO_NUM_21,                     // SDA pin
+                       false,                           // SDA internal pullup NOT enabled
+                       100 * 1000)                      // clock frequency - 100kHz
     {
     }
 
@@ -86,10 +68,7 @@ namespace test_multiple_4_line_devices
     {
         Application::init();
 
-        ili9341_initialized = init_ILI9341();
-        Log::info(TAG, "ILI9341 intialization --- {}", ili9341_initialized ? "Succeeded" : "Failed");
-
-        bme280_initialized = init_BME280SPI();
+        bme280_initialized = init_BME280I2C();
         Log::info(TAG, "BME280 intialization --- {}", bme280_initialized ? "Succeeded" : "Failed");
     }
 
@@ -99,80 +78,23 @@ namespace test_multiple_4_line_devices
         SystemStatistics::instance().dump();
         Log::info(TAG, "........................................" );
 
-        if (ili9341_initialized)
-        {
-            //print_display_parameters(READ_ID_INFO, THREE_PARAMS);
-            print_display_parameters(READ_DISPLAY_STATUS, FOUR_PARAMS);
-
-            //print_display_parameters(READ_DISPLAY_PWR_MODE, ONE_PARAM);
-            //print_display_parameters(READ_DISPLAY_MADCTL, ONE_PARAM);
-        }
-
         if (bme280_initialized)
         {
             print_thp_sensor_measurements();
             print_thp_sensor_configuration();
-            print_thp_sensor_id();
+
+            //print_thp_sensor_id();
         }
     }
 
-    bool App::init_ILI9341()
-    {
-        auto device = spi_master.create_device<ILI9341>(
-                        GPIO_NUM_14,            // chip select gpio pin
-                        GPIO_NUM_27,            // data command gpio pin
-                        GPIO_NUM_33,            // reset gpio pin
-                        GPIO_NUM_32,            // backlight gpio pin
-                        0,                      // spi command_bits
-                        0,                      // spi address_bits,
-                        0,                      // bits_between_address_and_data_phase,
-                        0,                      // spi_mode = 0,
-                        128,                    // spi positive_duty_cycle,
-                        0,                      // spi cs_ena_posttrans,
-                        SPI_MASTER_FREQ_16M,    // spi-sck = 16MHz
-                        0,                      // full duplex (4-wire)
-                        7,                      // queue_size,
-                        true,                   // use pre-trans callback
-                        true);                  // use post-trans callback
-
-        Log::info(TAG, "Initializing of SPI Device: ILI9341");
-        bool res = device->init(spi_host);
-
-        if (res)
-        {
-            device->reset_display();
-            res &= device->send_init_cmds();
-            device->set_back_light(true);
-            display = std::move(device);
-        }
-        else
-        {
-            Log::error(TAG, "Initializing of SPI Device: ILI9341 --- FAILED");
-        }
-
-        return res;
-    }
-
-    bool App::init_BME280SPI()
+    bool App::init_BME280I2C()
     {
         bool res = false;
-        auto device = spi_master.create_device<BME280SPI>(
-                        GPIO_NUM_13,            // chip select gpio pin
-                        0,                      // spi command_bits
-                        0,                      // spi address_bits,
-                        0,                      // bits_between_address_and_data_phase,
-                        0,                      // spi_mode = 0,
-                        128,                    // spi positive_duty_cycle,
-                        0,                      // spi cs_ena_posttrans,
-                        SPI_MASTER_FREQ_10M,    // spi-sck = 10MHz
-                        0,                      // full duplex (4-wire)
-                        2,                      // queue_size,
-                        false,                  // will not use pre-trans callback
-                        false);                 // will not use post-trans callback
+        auto device = i2c_master.create_device<BME280>(0x76);   // BME280 i2c address
 
-        Log::info(TAG, "Initializing of SPI Device: BME280SPI");
+        Log::info(TAG, "Scanning for BME280");
 
-        if (device->init(spi_host))
+        if (device->is_present())
         {
             Log::info(TAG, "BME280 reset: {}", device->reset());
 
@@ -204,25 +126,12 @@ namespace test_multiple_4_line_devices
         }
         else
         {
-            Log::error(TAG, "Initializing of SPI Device: BME280SPI --- FAILED");
+            Log::error(TAG, "BME280 not present");
         }
 
         thp_sensor = std::move(device);
 
         return res;
-    }
-
-    void App::print_display_parameters(const uint8_t cmd, uint8_t param_count)
-    {
-        std::vector<uint8_t> rxdata;
-        display->read_params(cmd, rxdata, param_count);
-
-        for (uint8_t i = 0; i < param_count; i++)
-        {
-            Log::error(TAG, "cmd = {:#04x}  param [{}] = {:#04x}", cmd, i, rxdata.at(i));
-        }
-
-        Log::info(TAG, "........................................" );
     }
 
     void App::print_thp_sensor_measurements()
@@ -233,7 +142,7 @@ namespace test_multiple_4_line_devices
         Log::error(TAG, "BME280 Temperature  (degC)  = {}", temperature);
         Log::error(TAG, "BME280 Humidity     (%RH)   = {}", humidity);
         Log::error(TAG, "BME280 Pressure     (hPa)   = {}", pressure / 100);
-        Log::error(TAG, "Barometric Pressure (in Hg) = {}", pressure / (float)3386.389);
+        Log::error(TAG, "Barometric Pressure (in Hg) = {}", pressure / static_cast<float> (3386.389));
         Log::info(TAG, "........................................" );
     }
