+++ test/spi_4_line_devices_test/CMakeLists.txt
